requirements.txt
flask
langchain
langchain-openai
langgraph
supabase
python-dotenv